+++ lib/digests/ripemd256.dart
@@ -69,7 +69,7 @@ class RIPEMD256Digest extends MD4FamilyDigest implements Digest {
     }
   }
 
-  void processLength(int bitLength) {
+  void processLength(Uint64 bitLength) {
     if (_xOff > 14) {
         processBlock();
     }
