+++ lib/digests/ripemd320.dart
@@ -73,7 +73,7 @@ class RIPEMD320Digest extends MD4FamilyDigest implements Digest {
     }
   }
 
-  void processLength(int bitLength) {
+  void processLength(Uint64 bitLength) {
     if (_xOff > 14) {
         processBlock();
     }
